generic linearizable and gaussian factors (#75)
This adds refined factor protocols "GenericLinearizableFactor" and "GenericGaussianFactor".

The main point is to allow solvers that do iterated linearization. The solver linearizes all the "GenericLinearizableFactors" at the current point to get a bunch of "GenericGaussianFactors", and then it solves the factor graph of "GenericGaussianFactors". (No solver included in this PR).